refactor: CartItem 기능 custom hook으로 관심사 분리
